renamed portal to be both origin and destination, added some basic styling for dev
